chore(retranspile): client 重生成归零——a2r_std time i64 对齐（Plan 396 §2.6）
auto-lang 283990bc：a2r_std::time::now_ms/now_sec 由滞后手抄的 i32 恢复为
stdlib/auto/time.rs.at 声明的 i64（epoch 毫秒 i32 截断每 ~24.8 天回绕，
本就是潜伏 bug）。daemon.rs 两个 deadline 随重生成变为 i64，
retranspile check 0 错——三转译 crate（ai-config/agent/client）首次同时全绿。

+++ crates/auto-ai-client/rust/src/daemon.rs
@@ -46,7 +46,7 @@ pub fn ensure_daemon() -> Option<String> {
             }
             
 
-            let deadline: i32 = a2r_std::time::now_ms() + 3000;
+            let deadline: i64 = a2r_std::time::now_ms() + 3000;
             while a2r_std::time::now_ms() < deadline {
                 a2r_std::time::sleep_ms(100 as u64);
                 if is_running() {
@@ -70,7 +70,7 @@ pub fn ensure_daemon_async() -> Option<String> {
             if spawn_daemon(aaid_path.as_str()) == false {
                 return None;
             }
-            let deadline: i32 = a2r_std::time::now_ms() + 10000;
+            let deadline: i64 = a2r_std::time::now_ms() + 10000;
             loop {
                 if is_running_async() {
                     return Some(daemon_url());
